Clean this up and document it.
