+++ src/util/Views.h
@@ -384,175 +384,3 @@ inline constexpr bool
     std::ranges::enable_borrowed_range<ad_utility::OwningView<T>> =
         std::ranges::enable_borrowed_range<T>;
 #endif
-
-namespace ad_utility {
-// This class is similar to `ql::views::transform` with the following
-// differences:
-// 1. The new values are computed when the iterators are advanced, not when they
-// are dereferenced.
-// 2. The computed values are therefore cached (using the `InputRangeFromGet`
-// facility), which might
-//    have a small performance overhead, as results that are used only once have
-//    to be moved and can't benefit from return value optimization (RVO).
-// 3. This class only yields an input range, independent of the range category
-// of the input.
-CPP_template(typename View, typename F)(
-    requires ql::ranges::input_range<View> CPP_and ql::ranges::view<View>
-        CPP_and std::is_object_v<F>)  // TODO requires clause for `F is callabel
-                                      // with the range reference  of `View` and
-                                      // is an object.
-    struct CachingTransformInputRange
-    : InputRangeFromGet<std::decay_t<
-          std::invoke_result_t<F, ql::ranges::range_reference_t<View>>>> {
- private:
-  View view_;
-  F f_;
-  using It = ql::ranges::iterator_t<View>;
-  using Res = std::decay_t<
-      std::invoke_result_t<F, ql::ranges::range_reference_t<View>>>;
-  std::optional<It> it_;
-
- public:
-  explicit CachingTransformInputRange(View view, F f = {})
-      : view_{std::move(view)}, f_(std::move(f)) {}
-
- private:
-  std::optional<Res> get() {
-    if (!it_.has_value()) {
-      it_ = view_.begin();
-    } else {
-      ++(it_.value());
-    }
-    if (*it_ == view_.end()) {
-      return std::nullopt;
-    }
-    return std::invoke(f_, *it_.value());
-  }
-  friend class InputRangeFromGet<Res>;
-};
-
-template <typename Range, typename F>
-CachingTransformInputRange(Range&&, F)
-    -> CachingTransformInputRange<all_t<Range>, F>;
-
-// Classes and types to represent control flows in generator-like statemachines,
-// like `break`, `continue`, and `return`.
-namespace loopControl {
-struct Continue {};
-struct Break {};
-template <typename T>
-struct BreakWithValue {
-  T value_;
-};
-
-// The central class to use the control flow mechanics.
-// For now, see the examples in the tests for the usage.
-// TODO<joka921> Write more documentation.
-template <typename T>
-struct LoopControl {
-  using type = T;
-  using V = std::variant<T, Continue, Break, BreakWithValue<T>>;
-  V v_;
-  bool isContinue() const { return std::holds_alternative<Continue>(v_); }
-  bool isBreak() const {
-    return std::holds_alternative<Break>(v_) ||
-           std::holds_alternative<BreakWithValue<T>>(v_);
-  }
-
-  std::optional<T> moveValueIfPresent() {
-    auto visitor = [](auto& v) -> std::optional<T> {
-      using U = std::remove_reference_t<decltype(v)>;
-      if constexpr (ad_utility::SimilarToAny<U, Continue, Break>) {
-        return std::nullopt;
-      } else if constexpr (ad_utility::SimilarTo<U, T>) {
-        return std::move(v);
-      } else {
-        static_assert(ad_utility::SimilarTo<U, BreakWithValue<T>>);
-        return std::move(v.value_);
-      }
-    };
-    return std::visit(visitor, v_);
-  }
-};
-
-template <typename T>
-struct LoopControlValue {};
-
-template <typename T>
-struct LoopControlValue<LoopControl<T>> {
-  using type = typename LoopControl<T>::type;
-};
-
-template <typename T>
-using loopControlValueT = typename LoopControlValue<T>::type;
-
-}  // namespace loopControl
-
-// This class can be used to simulate a generator that consists of a single
-// for-loop, but where the loop contains control-flows such as `break` and
-// `continue`. It is given an `View` and a functor `F` which must return a
-// `LoopControl` object (see above) when being called with the values of the
-// `View` with the following semantics:
-// 1. In principle the functor `F` is called for each element in the input
-// (unless there is a `break`, see below).
-// 2. If the value of `F(curElement)` is a `Continue` object, then the current
-// element is skipped, and nothing is yielded.
-// 3. If the value of `F(curElement)` is a `Break` object, then nothing is
-// yielded in the current step, and the iteration stops completely.
-// 4. If the value of `F(curElement)` is a `BreakWithValue` object, then the
-// value is yielded, but the iteration stops afterward (this is similar to a
-// `return` statement in a for-loop).
-// 5. Otherwise, the value of `F(curElement)` is a plain value, then that value
-// is yielded and the iteration resumes.
-
-CPP_template(typename View, typename F,
-             typename Res = loopControl::loopControlValueT<
-                 std::invoke_result_t<F, ql::ranges::range_reference_t<View>>>)(
-    requires ql::ranges::input_range<View> CPP_and ql::ranges::view<View>
-        CPP_and std::is_object_v<F>)  // TODO requires clause for `F is
-                                      // with the range reference  of `View` and
-                                      // returns a `LoopControl`.
-    struct CachingContinuableTransformInputRange : InputRangeFromGet<Res> {
- private:
-  View view_;
-  F f_;
-  bool receivedBreak_ = false;
-  using It = ql::ranges::iterator_t<View>;
-  std::optional<It> it_;
-
- public:
-  explicit CachingContinuableTransformInputRange(View view, F f = {})
-      : view_{std::move(view)}, f_(std::move(f)) {}
-
- private:
-  std::optional<Res> get() {
-    if (receivedBreak_) {
-      return std::nullopt;
-    }
-    while (true) {
-      if (!it_.has_value()) {
-        it_ = view_.begin();
-      } else {
-        ++(it_.value());
-      }
-      if (*it_ == view_.end()) {
-        return std::nullopt;
-      }
-      loopControl::LoopControl<Res> v = std::invoke(f_, *it_.value());
-      if (v.isContinue()) {
-        continue;
-      }
-      if (v.isBreak()) {
-        receivedBreak_ = true;
-      }
-      return v.moveValueIfPresent();
-    }
-  }
-  friend class InputRangeFromGet<Res>;
-};
-
-template <typename Range, typename F>
-CachingContinuableTransformInputRange(Range&&, F)
-    -> CachingContinuableTransformInputRange<all_t<Range>, F>;
-
-}  // namespace ad_utility
